修复数据类型不统一，删除冗余代码

+++ io/command.hpp
@@ -7,8 +7,8 @@ struct Command
 {
   bool control;
   bool shoot;
-  float yaw;
-  float pitch;
+  double yaw;
+  double pitch;
   double horizon_distance = 0;  //无人机专有
 };
